requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
joblib
pyyaml
#-e .